Separate PyTorch nodes
