+++ backend/src/nodes/nodes/onnx/save_model.py
@@ -1,0 +1,38 @@
+from __future__ import annotations
+
+import os
+
+from sanic.log import logger
+
+from ...categories import ONNXCategory
+from ...node_base import NodeBase
+from ...node_factory import NodeFactory
+from ...properties.inputs import OnnxModelInput, DirectoryInput, TextInput
+from ...utils.onnx_model import OnnxModel
+
+
+@NodeFactory.register("chainner:onnx:save_model")
+class OnnxSaveModelNode(NodeBase):
+    """ONNX Save Model node"""
+
+    def __init__(self):
+        super().__init__()
+        self.description = """Save ONNX model to file (.onnx)."""
+        self.inputs = [
+            OnnxModelInput(),
+            DirectoryInput(has_handle=True),
+            TextInput("Model Name"),
+        ]
+        self.outputs = []
+        self.category = ONNXCategory
+        self.name = "Save Model"
+        self.icon = "MdSave"
+        self.sub = "Input & Output"
+
+        self.side_effects = True
+
+    def run(self, model: OnnxModel, directory: str, model_name: str) -> None:
+        full_path = f"{os.path.join(directory, model_name)}.onnx"
+        logger.info(f"Writing file to path: {full_path}")
+        with open(full_path, "wb") as f:
+            f.write(model.bytes)
